TIC-TAC-TOE revised version
I changed variables, created new functions to make the code more readable and organized. There were a few errors with each of the functions, and their logic, and  I still need to write a function to check who won, and overall improve my code, so that it is a more complete version of the game.